Rename modeling module
