+++ src/script/execution/types/modeling/polygon.rs
